backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3==1.34.34",
    "fastapi==0.109.2",
    "httpx==0.27.0",
    "mangum==0.17.0",
    "moto[dynamodb,s3,ses]==5.0.1",
    "passlib[bcrypt]==1.7.4",
    "pydantic==2.6.1",
    "pydantic-settings==2.1.0",
    "pytest==8.0.0",
    "pytest-asyncio>=0.23.4",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.9",
    "qrcode[pil]==7.4.2",
    "reportlab==4.1.0",
    "twilio==9.0.0",
    "uvicorn==0.27.1",
]
